Applies the Single Responsibility Principle to output writer classes.
Before this commit, output writer files were responsible for calling
WayCalculator.calculate() to get results, and CurvatureRunner.js
was responsible for keeping it all together.

After this commit, OutputService is given the objects to output, and
a the configuration settings to determine which outputters to use.
App.js, is responsible for acting like the top level method it is.